feat: manifest

+++ lsm/table.go
@@ -28,7 +28,7 @@ import (
 type table struct {
 	ss   *file.SSTable
 	lm   *levelManager
-	fid  uint32
+	fid  uint64
 	idxs []byte
 }
