+++ utils/const.go
@@ -35,7 +35,7 @@ const (
 // codec
 var (
 	MagicText    = [4]byte{'H', 'A', 'R', 'D'}
-	MagicVersion = 1
+	MagicVersion = uint32(1)
 	// CastagnoliCrcTable is a CRC32 polynomial table
 	CastagnoliCrcTable = crc32.MakeTable(crc32.Castagnoli)
 )
